refactor [travel-package-controller]: removed getAllDestination and createDestination, then move it to destination.controller.js | added updateDestination
