+++ src/controllers/travel-package.controller.js
@@ -169,77 +169,6 @@ const createBundledTravelPackage = async (req, res) => {
   }
 };
 
-const createDestinations = async (req, res, _next) => {
-  uploadThumbnail(req, res, async (err) => {
-    if (err) {
-      return res.status(500).send({
-        success: false,
-        message: err.message,
-        data: null,
-      });
-    }
-
-    const { title, description } = req.body;
-    const thumbnail = req.file
-      ? `/uploads/thumbnails/${req.file.filename}`
-      : null;
-
-    if (!title) {
-      return res.status(400).send({
-        success: false,
-        message: "Title is required",
-        data: null,
-      });
-    }
-
-    try {
-      const newDestination = await Destinations.create({
-        title,
-        description,
-        thumbnail,
-      });
-
-      return res.status(200).send({
-        success: true,
-        message: "Destination created successfully",
-        data: newDestination,
-      });
-    } catch (error) {
-      return res.status(500).send({
-        success: false,
-        message: error.message,
-        data: null,
-      });
-    }
-  });
-};
-
-const getAllDestinations = async (req, res, _next) => {
-  try {
-    const destinations = await Destinations.findAll();
-
-    if (destinations.length === 0) {
-      return res.status(404).send({
-        success: false,
-        message: "No destinations found",
-        data: null,
-      });
-    }
-
-    return res.status(200).send({
-      success: true,
-      message: "Destinations retrieved successfully",
-      data: destinations,
-    });
-  } catch (error) {
-    return res.status(500).send({
-      success: false,
-      message: error.message,
-      data: null,
-    });
-  }
-};
-
 const createTravelPackage = async (req, res) => {
   try {
     uploadThumbnail(req, res, async (err) => {
@@ -329,8 +258,6 @@ const createTravelPackage = async (req, res) => {
 module.exports = {
   getAllTravelPackages,
   getTravelPackageById,
-  createDestinations,
   createTravelPackage,
   createBundledTravelPackage,
-  getAllDestinations,
 };
